requirements.txt
requests
bs4
tqdm
pyyaml